% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{adjacent_vertices}
\alias{adjacent_vertices}
\title{Adjacent vertices for all vertices in a graph}
\usage{
adjacent_vertices(graph)
}
\arguments{
\item{graph}{The graph.}
}
\value{
A named list of character vectors, the adjacent vertices
  for each vertex.
}
\description{
A vertex is adjacent is it is either a successor, or a predecessor.
}

