% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfs.R
\name{bfs}
\alias{bfs}
\title{Breadth-first search of a graph}
\usage{
bfs(graph, from = vertices(graph))
}
\arguments{
\item{graph}{Input graph.}

\item{from}{Character vector, which vertices to start the search
from. By default all vertices are attempted.}
}
\value{
Character vector of the named of the visited vertices,
  in the order of their visit.
}
\description{
Breadth-first search of a graph
}

