% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{degree}
\alias{degree}
\title{Degree of vertices}
\usage{
degree(graph, mode = c("out", "in", "total", "all"))
}
\arguments{
\item{graph}{\assert{is_simplegraph} Input graph.}

\item{mode}{Whether to calculate \code{out}-degree, \code{in}-degree,
or the \code{total} degree.}
}
\value{
Named numeric vector of degrees.
}
\description{
Degree of vertices
}
\examples{
G <- graph(list(A = c("B", "C"), B = "C", C = "A"))
degree(G, mode = "out")
degree(G, mode = "in")
degree(G, mode = "total")
}

