% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{incident_edges}
\alias{incident_edges}
\title{Incident edges}
\usage{
incident_edges(graph, mode = c("out", "in", "all", "total"))
}
\arguments{
\item{graph}{\assert{is_simplegraph} Input graph.}

\item{mode}{Whether to use \code{out} edges, \code{in} edges or
\code{all} edges.}
}
\value{
A list of data frames, each a set of edges.
}
\description{
Incident edges
}
\examples{
G <- graph(list(A = c("B", "C"), B = "C", C = "A"))
incident_edges(G, mode = "out")
incident_edges(G, mode = "in")
incident_edges(G, mode = "all")
}

