% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigraphs.R
\name{is_simple}
\alias{is_simple}
\title{Is this a simple graph?}
\usage{
is_simple(graph)
}
\arguments{
\item{graph}{The input graph.}
}
\value{
Logical scalar.
}
\description{
A simple graph contains no loop and multiple edges.
}
\seealso{
Other multigraphs: \code{\link{is_loopy}},
  \code{\link{is_multigraph}}, \code{\link{remove_loops}},
  \code{\link{remove_multiple}}, \code{\link{simplify}}
}

