% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_simplegraph}
\alias{is_simplegraph}
\title{Check if object is a simplegraph}
\usage{
is_simplegraph(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
TRUE if \code{x} is a \code{simplegraph} object. \code{FALSE}
  otherwise.
}
\description{
This is currently used for internal assertions
}
\keyword{internal}

