% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_vertices_of}
\alias{is_vertices_of}
\title{Check if the an object is a sequence of vertices from a graph}
\usage{
is_vertices_of(v, g)
}
\arguments{
\item{v}{R object.}

\item{g}{Graph.}
}
\value{
\code{TRUE} if \code{v} are vertices from \code{g}.
}
\description{
Check if the an object is a sequence of vertices from a graph
}
\keyword{internal}

