% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_named_lists}
\alias{merge_named_lists}
\title{Merge two names lists}
\usage{
merge_named_lists(list1, list2, default = character())
}
\arguments{
\item{list1}{First list.}

\item{list2}{Second list.}
}
\description{
Elements with common names are concatenated. If both lists have
\code{NULL} for an element, then \code{default} is used instead.
}
\keyword{internal}

