% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{predecessors}
\alias{predecessors}
\alias{successors}
\title{Predecessors and successors}
\usage{
predecessors(graph)

successors(graph)
}
\arguments{
\item{graph}{\assert{is_simplegraph} Input graph}
}
\value{
Named list of character vectors, the predecessors or
  the successors of each vertex.
}
\description{
Predecessors and successors
}
\examples{
G <- graph(list(A = c("B", "C"), B = "C", C = "A"))
predecessors(G)
successors(G)
}

