% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{sanitize}
\alias{sanitize}
\title{Check the validity of a graph data structure}
\usage{
sanitize(x, ...)
}
\arguments{
\item{x}{\assert{is_simplegraph} Graph.}

\item{...}{Extra arguments are curently ignored.}
}
\description{
This is mainly for internal checks, but occasionally it
might be useful externally.
}
\examples{
G <- graph(list(A = c("B", "C"), B = "C", C = "A"))
sanitize(G)

G <- c(G, list("this is not good" = c(1, 2, 3)))
try(sanitize(G))
}

