% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{simplegraph}
\alias{simplegraph}
\alias{simplegraph-package}
\title{Simple Graph Data Types and Basic Algorithms}
\description{
Simple classic graph algorithms for simple graph classes.
Graphs may possess vertex and edge attributes. 'simplegraph' has
no dependencies and it is writting entirely in R, so it is easy to
install.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gaborcsardi/simplegraph}
  \item Report bugs at \url{https://github.com/gaborcsardi/simplegraph/issues}
}

}
