% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{size}
\alias{size}
\title{The size of the graph is the number of edges}
\usage{
size(graph)
}
\arguments{
\item{graph}{The graph.}
}
\value{
Numeric scalar, the number of edges.
}
\description{
The size of the graph is the number of edges
}
\examples{
G <- graph(list(A = c("B", "C"), B = "C", C = "A"))
size(G)
}

