% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo-sort.R
\name{topological_sort}
\alias{topological_sort}
\title{Topological sorting of a graph}
\usage{
topological_sort(graph)
}
\arguments{
\item{graph}{\assert{is_simplegraph} Input graph.}
}
\value{
Character vector of vertex ids, in topological order.
}
\description{
Topological sorting of a graph
}
\examples{
funcs <- graph(list(
  drop_internal = character(0),
  get_deps = c("get_description", "parse_deps",
    "\%||\%", "drop_internal"),
  get_description = "pkg_from_filename",
  parse_deps = "str_trim",
  cran_file = c("get_pkg_type", "r_minor_version", "cran_file"),
  download_urls = c("split_pkg_names_versions", "cran_file"),
  filename_from_url = character(0),
  get_pkg_type = character(0),
  pkg_download = c("dir_exists", "download_urls",
    "filename_from_url", "try_download"),
  r_minor_version = character(0),
  try_download = character(0),
  drop_missing_deps = character(0),
  install_order = character(0),
  restore = c("pkg_download", "drop_missing_deps",
    "install_order", "get_deps"),
  snap = character(0),
  `\%||\%` = character(0),
  data_frame = character(0),
  dir_exists = character(0),
  pkg_from_filename = character(0),
  split_pkg_names_versions = "data_frame",
  str_trim = character(0)
))
topological_sort(remove_loops(funcs))
}

