% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tranpose.R
\name{transpose}
\alias{transpose}
\title{Transpose a graph}
\usage{
transpose(graph)
}
\arguments{
\item{graph}{Input graph}
}
\value{
Transposed graph.
}
\description{
The transposed graph have the same vertices, and the same number
of edges, but all edge directions are opposite comparated to the
original graph.
}

