% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{edges}
\alias{edges}
\title{Edges of a graph}
\usage{
edges(graph)
}
\arguments{
\item{graph}{The graph}
}
\value{
Data frame of edge data and metadata. The tail and head
  vertices are in the fist two columns. The rest of the columns are
  metadata.
}
\description{
Edges of a graph
}

