% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{graph}
\alias{graph}
\title{Create a graph}
\usage{
graph(x, ...)
}
\arguments{
\item{x}{A data frame, or a named list of character vectors. See
details below.}

\item{...}{Additional arguments, see details below.}
}
\value{
A graph object.
}
\description{
Graphs can be specified as adjacency lists or (two) data frames.
}
\details{
If the first argument is a data frame, then it is interpreted as
vertex data, and a second data frame must be supplied as edge data.
The first column of the vertex data must contain (character) vertex
ids. The first two columns of the edge data frame must contain the
directed edges of the graph, in the order of tail and head, as
characters referring to the nodes ids. Other columns are kept as
metadata.

If the first argument is not a data frame, but a list, then it is
interpreted as an adjacency list. It must be named, and the names
will be used as vertex ids. Each list element must be a character
vector containing the successors of each vertex.
}

