% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{order}
\alias{order}
\title{Order of a graph}
\usage{
order(graph)
}
\arguments{
\item{graph}{The graph.}
}
\value{
Numeric scalar, the number of vertices.
}
\description{
The order of the graph is the number of vertices.
}
\examples{
G <- graph(list(A = c("B", "C"), B = "C", C = "A"))
order(G)
}
\seealso{
Other simple queries: \code{\link{adjacent_vertices}},
  \code{\link{edges}}, \code{\link{vertices}}
}

