% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multigraphs.R
\name{simplify}
\alias{simplify}
\title{Remove multiple and loop edges from a graph}
\usage{
simplify(graph)
}
\arguments{
\item{graph}{\assert{is_simplegraph} Input graph.}
}
\value{
Another graph, with the multiple and loop edges removed.
}
\description{
Remove multiple and loop edges from a graph
}
\examples{
G <- graph(list(A = c("A", "B", "B"), B = c("A", "C"), C = "A"))
is_simple(G)

G2 <- simplify(G)
is_simple(G2)
}
\seealso{
Other multigraphs: \code{\link{is_loopy}},
  \code{\link{is_multigraph}}, \code{\link{is_simple}},
  \code{\link{remove_loops}}, \code{\link{remove_multiple}}
}

