% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topo-sort.R
\name{topological_sort}
\alias{topological_sort}
\title{Topological sorting of a graph}
\usage{
topological_sort(graph)
}
\arguments{
\item{graph}{Input graph.}
}
\value{
Character vector of vertex ids, in topological order.
}
\description{
Topological sorting of a graph
}

