% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{vertices}
\alias{vertices}
\title{Vertices of a graph, with metadata}
\usage{
vertices(graph)
}
\arguments{
\item{graph}{\assert{is_simplegraph} The graph.}
}
\value{
Character vector of vertex names.
}
\description{
Vertices of a graph, with metadata
}
\examples{
bridges <- graph(list(
  "Altstadt-Loebenicht" = c(
    "Kneiphof",
    "Kneiphof",
    "Lomse"
  ),
  "Kneiphof" = c(
    "Altstadt-Loebenicht",
    "Altstadt-Loebenicht",
    "Vorstadt-Haberberg",
    "Vorstadt-Haberberg",
    "Lomse"
  ),
  "Vorstadt-Haberberg" = c(
    "Kneiphof",
    "Kneiphof",
    "Lomse"
  ),
  "Lomse" = c(
    "Altstadt-Loebenicht",
    "Kneiphof",
    "Vorstadt-Haberberg"
  )
))
vertices(bridges)
}
\seealso{
Other simple queries: \code{\link{adjacent_vertices}},
  \code{\link{edges}}, \code{\link{order}}
}

